Команды редактирования категорий объединены в один файл
Лишние файлы удалены
